requirements.txt
numpy==1.18.5
scipy==1.5.0
kafka-python==2.0.2
termcolor==1.1.0
configparser==5.0.1
pandas==1.0.5
scikit_learn==0.23.2
prometheus-client==0.8.0
